+++ changelog.md
@@ -10,3 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a preliminary version of a data structure (playerMembersMap) to keep track of AM and PM turnip prices per guild member.
 - Added commands to wipe and list the playerMembersMap.
 - Added changelog!
+- Added DayData class for basic data structure to store AM and PM prices
+- Added amprice and pmprice commands
+
+### Changed
+- Deprecated price command. Related commands do not work anymore.
+- Changed help command printout to reflect commands that do not work with new amprice and pmprice commands.